test(sigil): cover required glyph selection changes
Implements T30
